Add files via upload
Given two models of equivalent "power", these visualizations demonstrate how the algorithm finds a path in the space of models which attempts to connect them while keeping the loss less than or equal to the starting models.

LongPolynomialConnection.gif is a very low error connection between two networks that perform cubic polynomial regression.

SymmetryDisconnect1 and 2 are two different runs of the same pair of models which fail to find a connection.  This is the counterexample 2-3-2 construction which we can prove has disconnections.  The starting pair of models is one model of with relatively low error and an identical copy of that model with two hidden nodes permuted.